creates sdk api wrapper, migrates leaderboard
